Move enqueueLinks() to its own file
